pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Harry5174 <harisjaved010@gmail.com>"]
readme = "README.md"
packages = [{include = "app", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = {extras = ["standard"], version = "^0.32.1"}
fastapi = {extras = ["standard"], version = "^0.115.6"}
sqlmodel = "^0.0.22"
python-dotenv = "^1.0.1"
asyncpg = "^0.30.0"
streamlit = "^1.40.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
